add util functions

+++ deq/deq_runtime/src/misc/index.rs
@@ -2,8 +2,8 @@ pub const WILDCARD: u64 = 0;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ErrorIndex {
-    pub eid: u64,
-    pub error_index: u64,
+    pub eid: usize,
+    pub error_index: usize,
 }
 
 // let the coordinator keep the internal loaded library
